khwtest: xlate_dev_mem_ptr is not exported from the kernel.
In kernel commit <e045fb2a988a9a1964059b0d33dbaf18d12f925f>, xlate_dev_mem_ptr
was moved from a define in a header to a function that is not exported. This
commit was first released in 2.6.26.

While there still might be some aliasing issues (which was the reason for the
move), I'll just use the older __pa() function directly.

I also updated the readme so the .ko goes into a place where it will be picked
up by default.

Signed-off-by: Shaun Ruffell <sruffell@digium.com>